pyproject.toml
[project]
name = "wildlifeai"
version = "0.1.0"
description = "WildlifeAI is an open-source Python framework that turns raw camera trap images into structured wildlife intelligence using a configurable, end-to-end AI pipeline."
authors = [
    { name = "Dhruva", email = "dhruvavishwak9@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pillow (>=12.0.0,<13.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{ include = "wildlifeai", from = "src" }]

[tool.poetry.scripts]
wildlifeai = "wildlifeai.cli:main"
